requirements.txt
pandas==2.2.3
requests==2.32.3
python-dotenv==1.1.0
pyarrow==19.0.0
